transfer_dump_requirements.txt
functions-framework==3.5.0
flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
google-cloud-storage==2.14.0
pyyaml==6.0.1
